+++ servant-cloudflare-workers/src/Servant/Cloudflare/Workers/Generic.hs
@@ -12,8 +12,8 @@ module Servant.Cloudflare.Workers.Generic (
   genericServe,
   genericServeT,
   genericServeTWithContext,
-  genericServer,
-  genericServerT,
+  genericWorker,
+  genericWorkerT,
 ) where
 
 import Data.Kind (
@@ -28,34 +28,35 @@ import Servant.Cloudflare.Workers.Internal
 
 -- | Transform a record of routes into a WAI 'Application'.
 genericServe ::
-  forall routes.
-  ( HasWorker (ToServantApi routes) '[]
-  , GenericServant routes AsWorker
-  , Worker (ToServantApi routes) ~ ToServant routes AsWorker
+  forall e routes.
+  ( HasWorker e (ToServantApi routes) '[]
+  , GenericServant routes (AsWorker e)
+  , Worker e (ToServantApi routes) ~ ToServant routes (AsWorker e)
   ) =>
-  routes AsWorker ->
-  Application
-genericServe = serve (Proxy :: Proxy (ToServantApi routes)) . genericServer
+  routes (AsWorker e) ->
+  FetchHandler e
+genericServe = serve (Proxy @e) (Proxy :: Proxy (ToServantApi routes)) . genericWorker
 
 {- | Transform a record of routes with custom monad into a WAI 'Application',
   by providing a transformation to bring each handler back in the 'Handler'
   monad.
 -}
 genericServeT ::
-  forall (routes :: Type -> Type) (m :: Type -> Type).
-  ( GenericServant routes (AsWorkerT m)
+  forall e (routes :: Type -> Type) (m :: Type -> Type).
+  ( GenericServant routes (AsWorkerT e m)
   , GenericServant routes AsApi
-  , HasWorker (ToServantApi routes) '[]
-  , WorkerT (ToServantApi routes) m ~ ToServant routes (AsWorkerT m)
+  , HasWorker e (ToServantApi routes) '[]
+  , WorkerT e (ToServantApi routes) m ~ ToServant routes (AsWorkerT e m)
   ) =>
   -- | 'hoistWorker' argument to come back to 'Handler'
-  (forall a. m a -> Handler a) ->
+  (forall a. m a -> Handler e a) ->
   -- | your record full of request handlers
-  routes (AsWorkerT m) ->
-  Application
-genericServeT f server = serve p $ hoistWorker p f (genericServerT server)
+  routes (AsWorkerT e m) ->
+  FetchHandler e
+genericServeT f worker = serve pe p $ hoistWorker pe p f (genericWorkerT worker)
   where
     p = genericApi (Proxy :: Proxy routes)
+    pe = Proxy @e
 
 {- | Transform a record of routes with custom monad into a WAI 'Application',
   while using the given 'Context' to serve the application (contexts are typically
@@ -63,41 +64,44 @@ genericServeT f server = serve p $ hoistWorker p f (genericServerT server)
   transformation to map all the handlers back to the 'Handler' monad.
 -}
 genericServeTWithContext ::
-  forall (routes :: Type -> Type) (m :: Type -> Type) (ctx :: [Type]).
-  ( GenericServant routes (AsWorkerT m)
+  forall e (routes :: Type -> Type) (m :: Type -> Type) (ctx :: [Type]).
+  ( GenericServant routes (AsWorkerT e m)
   , GenericServant routes AsApi
-  , HasWorker (ToServantApi routes) ctx
+  , HasWorker e (ToServantApi routes) ctx
   , HasContextEntry (ctx .++ DefaultErrorFormatters) ErrorFormatters
-  , WorkerT (ToServantApi routes) m ~ ToServant routes (AsWorkerT m)
+  , WorkerT e (ToServantApi routes) m ~ ToServant routes (AsWorkerT e m)
   ) =>
   -- | 'hoistWorker' argument to come back to 'Handler'
-  (forall a. m a -> Handler a) ->
+  (forall a. m a -> Handler e a) ->
   -- | your record full of request handlers
-  routes (AsWorkerT m) ->
+  routes (AsWorkerT e m) ->
   -- | the 'Context' to serve the application with
   Context ctx ->
-  Application
-genericServeTWithContext f server ctx =
-  serveWithContext p ctx $
-    hoistWorkerWithContext p pctx f (genericServerT server)
+  FetchHandler e
+genericServeTWithContext f worker ctx =
+  serveWithContext pe p ctx $
+    hoistWorkerWithContext pe p pctx f (genericWorkerT @e worker)
   where
+    pe = Proxy :: Proxy e
     p = genericApi (Proxy :: Proxy routes)
     pctx = Proxy :: Proxy ctx
 
 -- | Transform a record of endpoints into a 'Worker'.
-genericServer ::
-  (GenericServant routes AsWorker) =>
-  routes AsWorker ->
-  ToServant routes AsWorker
-genericServer = toServant
+genericWorker ::
+  forall e routes.
+  (GenericServant routes (AsWorker e)) =>
+  routes (AsWorker e) ->
+  ToServant routes (AsWorker e)
+genericWorker = toServant
 
 {- | Transform a record of endpoints into a @'WorkerT' m@.
 
  You can see an example usage of this function
  <https://docs.servant.dev/en/stable/cookbook/generic/Generic.html#using-generics-together-with-a-custom-monad in the Servant Cookbook>.
 -}
-genericServerT ::
-  (GenericServant routes (AsWorkerT m)) =>
-  routes (AsWorkerT m) ->
-  ToServant routes (AsWorkerT m)
-genericServerT = toServant
+genericWorkerT ::
+  forall e routes m.
+  (GenericServant routes (AsWorkerT e m)) =>
+  routes (AsWorkerT e m) ->
+  ToServant routes (AsWorkerT e m)
+genericWorkerT = toServant