+++ servant-cloudflare-workers/src/Servant/Cloudflare/Workers/Internal/Response.hs
@@ -3,20 +3,22 @@
 module Servant.Cloudflare.Workers.Internal.Response (
   PartialResponse (..),
   toWorkerResponse,
+  responseLBS,
 ) where
 
+import Control.Monad (guard)
 import Data.Aeson (Value)
 import qualified Data.ByteString as BS
+import qualified Data.ByteString.Lazy as LBS
 import qualified Data.CaseInsensitive as CI
 import qualified Data.Map.Strict as Map
 import Data.Word
 import GHC.Generics (Generic)
 import GHC.Wasm.Object.Builtins
 import GHC.Wasm.Web.JSON (encodeJSON)
-import Network.Cloudflare.Worker.Response (WorkerResponse, WorkerResponseBody, fromWorkerResponseBody, newResponse')
+import Network.Cloudflare.Worker.Response (WorkerResponse, WorkerResponseBody (..), fromWorkerResponseBody, newResponse')
 import qualified Network.Cloudflare.Worker.Response as Resp
-import Network.HTTP.Types (Status (..))
-import Network.HTTP.Types.Header (Header)
+import Network.HTTP.Types
 import qualified Wasm.Prelude.Linear as PL
 
 data PartialResponse = PartialResponse
@@ -44,3 +46,19 @@ toWorkerResponse PartialResponse {..} = do
         PL.. setPartialField "headers" (inject hdrs)
         PL.. setPartialField "encodeBody" encode
         PL.. setPartialField "cf" cf
+
+responseLBS ::
+  Status ->
+  [(HeaderName, BS.StrictByteString)] ->
+  LBS.ByteString ->
+  PartialResponse
+responseLBS status headers bdy =
+  PartialResponse
+    { status = status
+    , headers
+    , encodeBody = Nothing
+    , cloudflare = Nothing
+    , body = do
+        guard $ not $ LBS.null bdy
+        Just $ WorkerResponseLBS bdy
+    }